some fixes are performed
