+++ core/src/evaluation/functions/numeric/e.rs
@@ -1,0 +1,43 @@
+// Copyright 2024 The Drasi Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use async_trait::async_trait;
+use drasi_query_ast::ast;
+
+use crate::evaluation::functions::ScalarFunction;
+use crate::evaluation::variable_value::float::Float;
+use crate::evaluation::variable_value::VariableValue;
+use crate::evaluation::{ExpressionEvaluationContext, FunctionError, FunctionEvaluationError};
+
+#[derive(Debug)]
+pub struct E {}
+
+#[async_trait]
+impl ScalarFunction for E {
+    async fn call(
+        &self,
+        _context: &ExpressionEvaluationContext,
+        expression: &ast::FunctionExpression,
+        args: Vec<VariableValue>,
+    ) -> Result<VariableValue, FunctionError> {
+        if !args.is_empty() {
+            return Err(FunctionError {
+                function_name: expression.name.to_string(),
+                error: FunctionEvaluationError::InvalidArgumentCount,
+            });
+        }
+
+        Ok(VariableValue::Float(Float::from(std::f64::consts::E)))
+    }
+}
