+++ core/src/evaluation/functions/numeric/tests.rs
@@ -14,6 +14,7 @@
 
 mod abs_tests;
 mod ceil_tests;
+mod e_tests;
 mod exp_tests;
 mod floor_tests;
 mod random_tests;
